requirements.txt
praw
pandas
psycopg2-binary
python-dotenv
sqlalchemy